+++ src/encryption/c/aes/aes_cfb.h
@@ -7,14 +7,14 @@
 
 // Standard library-based implementations
 unsigned char* aes_cfb_encrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* aes_cfb_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
+unsigned char* aes_cfb_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 // Custom implementations
 unsigned char* aes_cfb_custom_encrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* aes_cfb_custom_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
+unsigned char* aes_cfb_custom_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 // OpenSSL-based implementations  
 unsigned char* aes_cfb_openssl_encrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* aes_cfb_openssl_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
+unsigned char* aes_cfb_openssl_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 #endif
