+++ src/encryption/c/aes/aes_common.h
@@ -15,10 +15,10 @@ void* aes_custom_init(void);
 void aes_custom_cleanup(void* context);
 
 // High-level encryption/decryption function that delegates to specific mode implementations
-unsigned char* aes_encrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
-unsigned char* aes_decrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
-unsigned char* aes_custom_encrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
-unsigned char* aes_custom_decrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
+unsigned char* aes_encrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
+unsigned char* aes_decrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
+unsigned char* aes_custom_encrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
+unsigned char* aes_custom_decrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
 
 // Stream processing functions
 unsigned char* aes_encrypt_stream(void* context, const unsigned char* data, int data_length, const unsigned char* key, int chunk_index, int* output_length);
