+++ src/encryption/c/aes/aes_ofb.c
@@ -69,7 +69,7 @@ unsigned char* aes_ofb_encrypt(aes_context_t* context, const unsigned char* data
     return output;
 }
 
-unsigned char* aes_ofb_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length) {
+unsigned char* aes_ofb_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length) {
     if (!context || !data || data_length <= 0) return NULL;
     
     // Calculate tag size for authentication
@@ -212,7 +212,7 @@ unsigned char* aes_ofb_custom_encrypt(aes_context_t* context, const unsigned cha
     return output;
 }
 
-unsigned char* aes_ofb_custom_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length) {
+unsigned char* aes_ofb_custom_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length) {
     if (!context || !data || data_length <= 0) return NULL;
     
     // Calculate tag size for authentication
@@ -382,7 +382,7 @@ unsigned char* aes_ofb_openssl_encrypt(aes_context_t* context, const unsigned ch
     return output;
 }
 
-unsigned char* aes_ofb_openssl_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length) {
+unsigned char* aes_ofb_openssl_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length) {
     if (!context || !data || data_length <= 0) return NULL;
     
     // Ensure we have enough data (at least for the IV)
@@ -480,7 +480,7 @@ unsigned char* aes_ofb_openssl_encrypt(aes_context_t* context, const unsigned ch
     return NULL;
 }
 
-unsigned char* aes_ofb_openssl_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length) {
+unsigned char* aes_ofb_openssl_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length) {
     fprintf(stderr, "Error: OpenSSL not available - AES-OFB OpenSSL implementation not supported\n");
     return NULL;
 }
