+++ src/encryption/c/aes/implementation.h
@@ -28,8 +28,8 @@ void register_aes_implementations(implementation_registry_t* registry);
 void* aes_init(void);
 void aes_cleanup(void* context);
 unsigned char* aes_generate_key(void* context, int* key_length);
-unsigned char* aes_encrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
-unsigned char* aes_decrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
+unsigned char* aes_encrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
+unsigned char* aes_decrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
 
 // Stream processing functions
 unsigned char* aes_encrypt_stream(void* context, const unsigned char* data, int data_length, const unsigned char* key, int chunk_index, int* output_length);
@@ -39,24 +39,24 @@ unsigned char* aes_decrypt_stream(void* context, const unsigned char* data, int 
 void* aes_custom_init(void);
 void aes_custom_cleanup(void* context);
 unsigned char* aes_custom_generate_key(void* context, int* key_length);
-unsigned char* aes_custom_encrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
-unsigned char* aes_custom_decrypt(void* context, const unsigned char* data, int data_length, const unsigned char* key, int* output_length);
+unsigned char* aes_custom_encrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
+unsigned char* aes_custom_decrypt(void* context, const unsigned char* data, size_t data_length, const unsigned char* key, size_t* output_length);
 
 // Mode-specific implementations
 // AES-CBC
 unsigned char* aes_cbc_encrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* aes_cbc_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
+unsigned char* aes_cbc_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 // AES-CTR
 unsigned char* aes_ctr_encrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* aes_ctr_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
+unsigned char* aes_ctr_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 // AES-GCM
-unsigned char* aes_gcm_encrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* aes_gcm_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
+unsigned char* aes_gcm_encrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
+unsigned char* aes_gcm_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 // AES-ECB
 unsigned char* aes_ecb_encrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* aes_ecb_decrypt(aes_context_t* context, const unsigned char* data, int data_length, int* output_length);
+unsigned char* aes_ecb_decrypt(aes_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 #endif /* AES_IMPLEMENTATION_H */
