+++ src/encryption/c/c_core.h
@@ -31,8 +31,8 @@ typedef struct {
     void* (*init)(void);
     void (*cleanup)(void*);
     unsigned char* (*generate_key)(void*, int*);
-    unsigned char* (*encrypt)(void*, const unsigned char*, int, const unsigned char*, int*);
-    unsigned char* (*decrypt)(void*, const unsigned char*, int, const unsigned char*, int*);
+    unsigned char* (*encrypt)(void*, const unsigned char*, size_t, const unsigned char*, size_t*);
+    unsigned char* (*decrypt)(void*, const unsigned char*, size_t, const unsigned char*, size_t*);
     
     // Stream processing functions
     unsigned char* (*encrypt_stream)(void*, const unsigned char*, int, const unsigned char*, int, int*);
